tidy ssh test build rules; move common field matching code to test_field_reflect.h
